move some git related functions to a git_utils specific helper module
